requirements/base.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
decorator==5.2.1
executing==2.2.0
frozenlist==1.7.0
humanfriendly==10.0
idna==3.10
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.6.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.1.1
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
stack-data==0.6.3
Telethon==1.40.0
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
yarl==1.20.1
